whisper/requirements.txt
# 反诈AI电话专员 - 依赖包列表

# 核心音频处理
sounddevice>=0.4.6
soundfile>=0.12.0
numpy>=1.21.0

# 语音识别 (Whisper)
faster-whisper>=1.0.0
torch>=2.0.0
torchaudio>=2.0.0
onnxruntime>=1.15.0

# TTS语音合成
edge-tts>=6.1.0
pyttsx3>=2.90

# 音频播放
pygame>=2.5.0
pyinstaller>=6.6.0
pillow>=10.0.0

# AI对话
openai>=1.0.0
python-dotenv>=1.0.0

# 工具库
requests>=2.28.0
httpx>=0.24.0

sounddevice